test: unify deadline error checking into testing.AssertIsTimeoutError
