+++ lib/testing/server.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package proto
+package testing
 
 import (
 	"errors"
